tools/agentsec/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agentsec"
version = "0.1.0"
description = "Validate and enforce AGENTSECURITY.md policies for autonomous AI agents"
readme = "README.md"
license = {text = "Apache-2.0"}
requires-python = ">=3.9"
authors = [
    {name = "AgentSecurity Project", email = "hello@agentsecurity.dev"}
]
keywords = ["ai", "agents", "security", "validation", "compliance", "owasp", "nist"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Security",
    "Topic :: Software Development :: Quality Assurance",
]
dependencies = [
    "pyyaml>=6.0",
    "jsonschema>=4.20",
    "click>=8.1",
    "rich>=13.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
]

[project.scripts]
agentsec = "agentsec.cli:main"

[project.urls]
Homepage = "https://agentsecurity.dev"
Repository = "https://github.com/agentsecurity/agentsecurity"
Documentation = "https://agentsecurity.dev/specification"
Issues = "https://github.com/agentsecurity/agentsecurity/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
